:lipstick: Add 본인 생년월일 등록화면 UI
커플연결 4/5_본인의 생년월일을 캘린더에서 선택하는 화면입니다.
Related to:#5
